Add support for @Swift(Weak) attribute on properties
Added support for `@Swift(Weak)` attribute on properties of interfaces.
Such properties are generated with `weak` keyword in Swift. Protocols
generated in Swift for iterfaces with such properties can only be
implemented by Swift code. CBridge infrastructure for backing such
interfaces with C++ objects is not generated.

Added SwiftWeakPropertiesValidator to validate the correct usage of the
new attribute. Added unit tests for the validator.

Added smoke and functional tests as appropriate.

Resolves: #465
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>